Refactor categoryController and associations***
This commit refactors the categoryController.js file by simplifying the getProductsByCategory function. It removes unnecessary code and replaces it with a single include statement for the Product model.

Additionally, the commit modifies the associations in the ProductCategories.js file. It enables the belongsToMany relationship between Product and Category, allowing products to have multiple categories and categories to have multiple products.